+++ Sources/Environment.swift
@@ -5,11 +5,12 @@ public struct Environment {
 
   public var loader: Loader?
 
-  public init(loader: Loader? = nil,
-              extensions: [Extension] = [],
-              templateClass: Template.Type = Template.self,
-              trimBehavior: TrimBehavior = .none) {
-
+  public init(
+    loader: Loader? = nil,
+    extensions: [Extension] = [],
+    templateClass: Template.Type = Template.self,
+    trimBehavior: TrimBehavior = .none
+  ) {
     self.templateClass = templateClass
     self.loader = loader
     self.extensions = extensions + [DefaultExtension()]
@@ -47,5 +48,4 @@ public struct Environment {
     template.environment = self
     return try template.render(context)
   }
-
 }