+++ Sources/Filters.swift
@@ -108,8 +108,8 @@ func indent(_ content: String, indentation: String, indentFirst: Bool) -> String
 
   var lines = content.components(separatedBy: .newlines)
   let firstLine = (indentFirst ? indentation : "") + lines.removeFirst()
-  let result = lines.reduce([firstLine]) { result, line in
-    result + [(line.isEmpty ? "" : "\(indentation)\(line)")]
+  let result = lines.reduce(into: [firstLine]) { result, line in
+    result.append(line.isEmpty ? "" : "\(indentation)\(line)")
   }
   return result.joined(separator: "\n")
 }